requirements.txt
# core blockchain interaction
web3==6.11.3
eth-abi==4.2.1
eth-account==0.9.0

# mathematical computation
numpy==1.24.3
scipy==1.11.2
pandas==2.0.3

# graph algorithms and optimization
networkx==3.1
cvxpy==1.4.1

# async and networking
aiohttp==3.8.5
websockets==11.0.3
asyncio-throttle==1.0.2

# data visualization
matplotlib==3.7.2
plotly==5.15.0
seaborn==0.12.2

# machine learning (optional)
scikit-learn==1.3.0
# tensorflow==2.13.0  # uncomment for ml experiments

# development tools
python-dotenv==1.0.0
rich==13.5.2
typer==0.9.0

# testing
pytest==7.4.2
pytest-asyncio==0.21.1
